pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = ["Alexis Chân Gridel <alexis.gridel@pm.me>"]
license = "GPL-3.0-only"
readme = "README.md"
package-mode = false  

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
discord-py = ">=2.4.0,<3.0.0"
asyncpg = "^0.30.0"
pydantic-settings = "^2.7.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.1"
black = "^24.10.0"
pre-commit = "^4.0.1"
watchfiles = "^1.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "Q", "I"]
ignore = ["E203", "E501"]

[tool.black]
line-length = 88